Move demo data to a separate package
