+++ demo/data/README.md
@@ -1,0 +1,5 @@
+# RainCatcher Demo Data
+
+This is a sample module containing static data definitions to seed an example RainCatcher project.
+
+It has methods to populate a mongo database with its data.
